Update our otel collector config generator to be compatible with otel v0.33.0 (#7)
Which problem is this PR solving?
Otel Collector was recently upgraded to v0.33.0 and included a number of breaking changes. In order to work with otlp v0.9 and the new data model changes, we must support Otel Collector v0.33.0

Short description of the changes
Changed all library dependencies to use v0.33.0.

Testing
Unit tests passed and tested locally.