bring back the python-installed yq
While go-yq *is* available in the CI runner, we need python-yq to
support defining functions in jq-programs and some other specific
syntax currently in use.

Update the bin/{jq|yq} wrappers to use local binaries on PATH if they
exist. In yq's case, also check that it is the python edition.